updates to registry keeper structure to isolate message passing implementations from keeper type
