+++ internal/blockchain/x/registry/msg/msg_server_access_application.go
@@ -1,4 +1,4 @@
-package keeper
+package msg
 
 import (
 	"context"
