+++ internal/blockchain/x/registry/msg/msg_server_test.go
@@ -1,4 +1,4 @@
-package keeper_test
+package msg_test
 
 import (
 	"context"
@@ -6,11 +6,11 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	keepertest "github.com/sonr-io/sonr/internal/blockchain/testutil/keeper"
-	"github.com/sonr-io/sonr/internal/blockchain/x/registry/keeper"
+	"github.com/sonr-io/sonr/internal/blockchain/x/registry/msg"
 	"github.com/sonr-io/sonr/internal/blockchain/x/registry/types"
 )
 
 func setupMsgServer(t testing.TB) (types.MsgServer, context.Context) {
 	k, ctx := keepertest.RegistryKeeper(t)
-	return keeper.NewMsgServerImpl(*k), sdk.WrapSDKContext(ctx)
+	return msg.NewMsgServerImpl(*k), sdk.WrapSDKContext(ctx)
 }
